xml: Automatically parse external entity references
PdbXml now automatically handles a stack of external reference parsers
so the upper layers will see the events as if its a single document.

The stack keeps track of the filenames which can now be directly
queried from the xml parser object.

The pdb_xml_stop_parser mechanism has been replaced with an abort
function that explicitly takes a GError object to pass up.