pdb-db: Just copy the XML instead of trying to convert to HTML
It looks like the WebView in Android doesn't work very well so it will
probably be easier to parse the XML in the application. prevodb now
mostly just copies the tags across except it still replaces the <tld>
references, expands entities, removes repeated whitespace and replaces
the marks with an integer.