Change the trie format to allow multiple articles for a keyword
When a valid word is encountered in a tree it is now followed by a
list of articles and marks instead of just one. Also an article can
have an optional alternative name to display instead of the one stored
in the index. That way it is possible to have "Britujo" be listed
under the index "britujo". PdbDb now always converts the index name to
lowercase before passing it on to the trie and will use the optional
display name if that caused the name to change.