prevodb: Resolve the article and section numbers later
Instead of directly storing the article and mark number in the index
when the reference is first encountered, the PdbDb now just stores a
pointer to the article and section. The numbers are calculated later
once all of the articles and sections have been generated. That way
the article and section doesn't need to be added to the arrays
immediately to know its number.

The PdbTrie now stores a data pointer for each word and has a callback
function to calculate the article and mark number from that data
pointer while the trie is being compressed.