pdb-db: Add spans for the <ofc> and <ekz> tags
The ofc tags are marked with a superscript span and the ekz tags are
marked with an italic span.

The offsets and lengths for the spans have been converted to represent
the length of the string if it was encoded in UTF-16. This makes it
easier to handle in the Java side.

When a span is started a closing marker is added to the processing
stack so that it knows to store the end of the span after all of the
children are processed.