prevodb: Handle <ref> tags and mrk attributes
Whenever a <drv> is parsed the child nodes are scanned for mrk
attributes and a mark is added for each.

The <ref> tags now cause a reference span to be added with the
corresponding PdbDbReference. The list of references is now stored
against the PdbDb instead of against each article because there is no
reason not to and it means the handler for the ref tag doesn't need a
pointer to the article.