Add styles for rim, em and aut tags
The <aut> tag needs to add [ and ] arround the span so there is now a
new stack entry type for adding a character after all of the children
have been processed.

The <rim> tag forms a paragraph and has a span. The span needs to be
applied after the paragraph separator. To make this easier there is
now a 'paragraph' field in the span descriptions to make a paragraph
around the span. pdb_db_parse_node makes sure to handle the paragraph
before the span is started.