pdb-db: Support variations for headwords
Some <kap> tags contain <var> tags for a variation of the word with
exactly the same meaning. The <var> tags were being ignored so only
the first word would appear with a load of commas after it. This patch
makes it strip the commas and recursively add the headword from the
embedded <var> tags to the index. The embedded <kap> tags are no
longer ignored if they have a parent <kap> tag.