prevodb: Add a tool for reading the single file database
This adds a tool for showing an article from the single file database.
It takes an optional language code and a word from the index and
displays the strings from the article. The idea is eventually to
format these strings either with nroff or html and display them in
some pager.

The tool can also complete language and word names with the intention
that eventually this could be integrated with bash completion to make
a convenient command line dictionary.