pdb: Pipe the output of the prevo tool to groff and less
The generated article for the prevo tool is now passed on to a bash
subprocess which sets up a pipe between groff and less. Groff is
configured to parse a man page. The article is still generated as
normal text but the idea is to change it be in nroff format.