Jump to the subsection for the selected mark
When the index points to a subsection of the article, the prevo tool
now inserts a marker into the generated man page at the end of the
title for the selected definition. An option is then set in the LESS
environment variable to make less search for the mark. That way it
will jump to the right definition.