Don't link prevo to libexpat
Only the prevodb tool and some of the test programs actually need to
link against expat. However, AS_CHECK_LIB was being used in the
configure script with no default action for if-found which makes it
put the library in LIBS. That meant that all of the programs would
link against it. This sets a dummy if-found argument and explicitly
adds -lexpat to the binaries that need it.