Add language codes to the sections in each article
Each section of the articles is now preceded by a language code which
is up to three letters long. Because the format has changed, the last
letter of the magic number for the single-file format has been
converted to be a version number. The version number is now 'C'
instead of 'B'. The unix client can read both formats.

The idea is that this will be used in the Android client to avoid
showing some translations depending on the user's preferences.