Group articles into files containing a set of 16
The package installer on Android 2.1 seems to have a problem with
packages containing a large number of files which causes it to crash
while installing. To work around this, the articles are now grouped
into files containing up to 16 articles. The articles are named using
the most significant 3 nibbles of the 16-bit article number. The
article file format now always includes the article length so that the
Android application can detect the end of the article.