Allow specifying multiple source database directories
The -i option can now be specified multiple times. When trying to open a
file, PdbRevo will now search for the file in all of the directories.
The main reason to do this is to make it easier to directly use the
official ReVo repos for the database. These are split into two repos,
one for the configuration files in voko-grundo and the other for the XML
files in revo-fonto.