xml: Manually handle .. in paths
When handling entities that begin with ‘..’, PdbXml now manually handles
removing parts of the base directory name. This makes it easier to have
the database split across multiple directories. For example, if a file
in the revo directory of the revo-fonto repo references
“../dtd/something.xml”, then without this patch it would try to look for
“revo/../dtd/something.xml” but it wouldn’t be found in the voko-grundo
repo because that has no “revo” directory.