db: Don’t add trailing spaces in spannable strings
Instead of directly adding a space when it is encountered in the XML, it
is now queued like it already does for paragraphs. That way if there are
only trailing spaces before the next paragraph or the end of the string
they will be skipped.

This should also fix a bug reported by Robin in the PReVo app where the
trailing spaces in the section headers cause the PIV search to break.

https://github.com/bpeel/prevo/pull/11